refactor/frontend/bulletproof architectureを採用
