+++ src/app/posts/[slug]/page.tsx
@@ -3,11 +3,11 @@ import { notFound } from "next/navigation";
 import { getAllPosts, getPostBySlug } from "@/lib/api";
 import { CMS_NAME } from "@/lib/constants";
 import markdownToHtml from "@/lib/markdownToHtml";
-import Alert from "@/app/_components/alert";
-import Container from "@/app/_components/container";
-import Header from "@/app/_components/header";
-import { PostBody } from "@/app/_components/post-body";
-import { PostHeader } from "@/app/_components/post-header";
+import Alert from "@/components/alert/Alert";
+import Container from "@/components/ui/Container";
+import Header from "@/components/layout/Header";
+import { PostBody } from "@/features/articles/components/PostBody";
+import { PostHeader } from "@/features/articles/components/PostHeader";
 
 export default async function Post(props: Params) {
   const params = await props.params;
