+++ src/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import Container from "@/app/_components/container";
+import Container from "@/components/ui/Container";
 
 export function Footer() {
   return (
