+++ src/components/theme/ThemeSwitcher.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import styles from "./switch.module.css";
+import styles from "@/components/ui/Switch.module.css";
 import { memo, useEffect, useState } from "react";
 
 declare global {
