+++ src/features/articles/components/CoverImage.tsx
